bug fix for being unable to import older CharTool rptoks; jfrazierjr_Individual_Views_and_GMPlayerView_Vision_Arcs.txt

+++ src/net/rptools/maptool/client/ui/ZoneImageGenerator.java
@@ -3,10 +3,6 @@
  */
 package net.rptools.maptool.client.ui;
 
-import net.rptools.maptool.client.ui.zone.PlayerView;
-import net.rptools.maptool.client.ui.zone.ZoneRenderer;
-
-import java.awt.Color;
 import java.awt.Graphics;
 import java.awt.Graphics2D;
 import java.awt.Point;
@@ -18,12 +14,15 @@ import java.awt.image.Raster;
 import java.awt.image.TileObserver;
 import java.awt.image.WritableRaster;
 
+import net.rptools.maptool.client.ui.zone.PlayerView;
+import net.rptools.maptool.client.ui.zone.ZoneRenderer;
+
 /**
  * This is a read-only implementation of BufferedImage intended to
  * be used with ImageWriters to create very large graphics
  * files. Since this is read-only, it gets its pixel data
  * by rasterizing a Zone piece-meal as the data is requested
- * by getData(rectangle) calls. 
+ * by getData(rectangle) calls.
  * 
  * A very simple 4 MB cache is implemented to reduce zoneRender calls.
  */
@@ -33,7 +32,7 @@ public class ZoneImageGenerator extends BufferedImage {
 	final ZoneRenderer renderer;
 	final PlayerView   view;
 	final Rectangle    origBounds;
-	
+
 	final int     maxCacheSize = 1024*1024; // probably 4MB for most systems
 	Raster        cachedRaster;
 	Rectangle     cachedRect;
@@ -53,26 +52,24 @@ public class ZoneImageGenerator extends BufferedImage {
 
 	private boolean cacheMiss(Rectangle rect) {
 		boolean miss;
-		
 		miss = (cachedRect == null) || !cachedRect.contains(rect);
-
 		return miss;
 	}
-	
+
 	@Override
 	public Raster getData(Rectangle rect) {
 		if (cacheMiss(rect)) {
 			// Figure out what slice of the Zone to cache
 			int sizeX = rect.width;
 			int sizeY = rect.height;
-			// Let's first try making the cache as wide as the whole zone 
+			// Let's first try making the cache as wide as the whole zone
 			sizeX = Math.max(sizeX, renderer.getBounds().width);
-			sizeY = Math.max(sizeY, (int) (maxCacheSize / sizeX));
+			sizeY = Math.max(sizeY, (maxCacheSize / sizeX));
 			if ((sizeX * sizeY) > maxCacheSize) {
-				// That didn't work, so let's try just making it as wide as requested 
-				// and as tall as possible 
+				// That didn't work, so let's try just making it as wide as requested
+				// and as tall as possible
 				sizeX = rect.width;
-				sizeY = (int) (maxCacheSize / sizeX);
+				sizeY = (maxCacheSize / sizeX);
 			}
 			// And let's make sure not to overdo things: no need for the cache to
 			// be larger than the zone itself!
@@ -87,29 +84,27 @@ public class ZoneImageGenerator extends BufferedImage {
 			numMisses++;
 			recentHits = 0;
 		}
-		
 		recentHits++;
 		return cachedRaster.createChild(
-				rect.x, rect.y,             // source upper-left X, Y 
+				rect.x, rect.y,             // source upper-left X, Y
 				rect.width, rect.height,    // size
-               	rect.x, rect.y,             // child upper-left X, Y (this is 'translated', not actual pixel coords. 
-               	                            // The returned raster is only as big as width * height
-               	null);
-
+				rect.x, rect.y,             // child upper-left X, Y (this is 'translated', not actual pixel coords.
+				// The returned raster is only as big as width * height
+				null);
 	}
-	
-	private void fillCache() { 
+
+	private void fillCache() {
 		Rectangle rect = cachedRect;
-		
+
 		if ((recentHits == 0) && (numMisses > 0)) {
 			assert false: "Cache is being thrashed: " + prevCacheRect + cachedRect;
 		}
-		
+
 		// preserve settings
 		Scale     origScale  = new Scale(renderer.getZoneScale());
 		Rectangle origBounds = new Rectangle(renderer.getBounds());
 
-		// set new temp vars 
+		// set new temp vars
 		Scale s = new Scale(origScale);
 		s.setOffset(
 				origScale.getOffsetX() - rect.x,
@@ -117,27 +112,26 @@ public class ZoneImageGenerator extends BufferedImage {
 		renderer.setZoneScale(s);
 		renderer.setBounds(rect);
 
-		// make a tiny buffered image for this (hopefully) small rectangle request 
+		// make a tiny buffered image for this (hopefully) small rectangle request
 		BufferedImage image = new BufferedImage(rect.width, rect.height, super.getType());
 		Graphics2D g = image.createGraphics();
 		g.setClip(0, 0, rect.width, rect.height);
 		renderer.renderZone(g, view);
 		// dispose is probably not needed. According to javadocs g's are disposed automatically when used in paint()
-		g.dispose(); 
+		g.dispose();
 		// makes a copy of the raster...
-		Raster raster = image.getData(); 
+		Raster raster = image.getData();
 		// ...so we can nudge it back to the original coordinate system
-		cachedRaster = raster.createTranslatedChild(rect.x, rect.y); 
+		cachedRaster = raster.createTranslatedChild(rect.x, rect.y);
 		image = null;
 		raster = null;
 
 		renderer.setBounds(origBounds);
 		renderer.setZoneScale(origScale);
-		
 	}
 
 	///////////////////////////////////////////////////////////////////////
-	// All of the methods after this are various forms of no-op designed 
+	// All of the methods after this are various forms of no-op designed
 	// to ensure the application fails in a predictable way
 	// if it tries to write to this object.
 	// As noted above, this is a read-only object!
@@ -172,7 +166,7 @@ public class ZoneImageGenerator extends BufferedImage {
 
 	/**
 	 *  To the outside world, we represent that this image is as large
-	 *  as the Component. 
+	 *  as the Component.
 	 */
 	@Override
 	public int getWidth() {
@@ -181,7 +175,7 @@ public class ZoneImageGenerator extends BufferedImage {
 
 	/**
 	 *  To the outside world, we represent that this image is as large
-	 *  as the Component. 
+	 *  as the Component.
 	 */
 	@Override
 	public int getHeight() {
@@ -190,7 +184,7 @@ public class ZoneImageGenerator extends BufferedImage {
 
 	/**
 	 *  To the outside world, we represent that this image is as large
-	 *  as the Component. 
+	 *  as the Component.
 	 */
 	@Override
 	public int getWidth(ImageObserver observer) {
@@ -199,7 +193,7 @@ public class ZoneImageGenerator extends BufferedImage {
 
 	/**
 	 *  To the outside world, we represent that this image is as large
-	 *  as the Component. 
+	 *  as the Component.
 	 */
 	@Override
 	public int getHeight(ImageObserver observer) {
@@ -225,7 +219,7 @@ public class ZoneImageGenerator extends BufferedImage {
 	@Override
 	public Raster getData() {
 		assert false: "Can not get a raster for the whole CachedComponentImage!";
-		return null;
+	return null;
 	}
 
 
@@ -252,13 +246,13 @@ public class ZoneImageGenerator extends BufferedImage {
 		// I would like to throw an exception, but I can't since BufferedImage doesn't
 	}
 
-	// These methods are 
+	// These methods are
 	//
 	//
 	@Override
 	public void removeTileObserver(TileObserver to) {
 		super.removeTileObserver(to);
-		// Ha! In BufferedImage image this is a no-op! 
+		// Ha! In BufferedImage image this is a no-op!
 	}
 
 	/**
@@ -301,5 +295,4 @@ public class ZoneImageGenerator extends BufferedImage {
 		// Do nothing, since this object is read-only.
 		// I would like to throw an exception, but I can't since BufferedImage doesn't
 	}
-
 }
